fix: resolve compiler warnings, add centralized message capping, and clean up code
- Remove unused imports (EcommerceChannel, LoyaltyProgram) from game/mod.rs
- Add GameState::push_message() with automatic message cap enforcement
- Replace all state.messages.push() in routes.rs with push_message() to prevent
  unbounded message growth from repeated UI interactions
- Use LoyaltyTier::icon() in routes instead of duplicate match block
- Fix clippy unnecessary cast (u32 -> u32) in loyalty.rs
- Tighten #[allow(dead_code)] to only cover truly unused PlayerActions fields
- Remove duplicate message cap loop from simulate_quarter (handled by push_message)
- Zero compiler warnings and zero clippy warnings

+++ src/game/loyalty.rs
@@ -180,7 +180,7 @@ pub fn update_loyalty(state: &mut GameState) -> f64 {
         .saturating_sub(churn)
         .saturating_add(new_members);
 
-    let member_cap = operating_count as u32 * 15_000;
+    let member_cap = operating_count * 15_000;
     program.members = program.members.min(member_cap);
 
     let cost = program.tier.quarterly_cost_per_store() * operating_count as f64;
